Gate implementation. (#1)
(1/4) Added some basic classes to gate.h and gate.cpp.

(2/4) Added basic functionality to Gate.

(3/4) Fixed some documentation.

(4/4) Changed Gate methods for convertSignal to return a QString representign the new ID that will be assigned to any wire that is conencted to the output side of this gate. This involved addign wire ID parameters. Updated documentation to reflect this change.